plugin: Must handle sample rate / buffer size changes properly
Must reinit synthesizer on every buffer size changes.
Otherwise Yoshimi will behave unexpectedly on VST3 and CLAP:

- Crash when destroying Parts (during destructor of SynthEngine)!
- Generate wrong samples (REAPER will automute the track)!

I've debugged for a long time, until it occured to me that buffer size
was not handled properly in my plugin.

So it is with sample rate: not properly handling sample rate changes
will render a terrible drill sound!